renamed and cleaned up main js
